blogger/requirements.txt
google-cloud-aiplatform==1.26.0
python-dotenv==1.0.0
ipython==8.14.0
google-api-python-client==2.90.0
google-auth-oauthlib==1.0.0